feat(cmd/goduck): delete the commands to start and stop application chains
1. Delete the commands to start and stop fabric chain;
2. Delete the commands to start and stop ethereum chain;
3. Adjust the position of some function to make the function of starting/stoping appchain in [appchain]/[appchain].go and make the function of interacting with appchain contract in [appchain]/contract.go.
