+++ cmd/goduck/goduck.go
@@ -33,8 +33,8 @@ func main() {
 		getVersionCMD(),
 		GetInitCMD(),
 		GetStatusCMD(),
-		fabric.GetFabricCMD(),
-		ethereum.GetEtherCMD(),
+		fabric.FabricContractCMD(),
+		ethereum.EtherContractCMD(),
 		keyCMD(),
 		bitxhubCMD(),
 		pierCMD,
